Split sidebar ExtensionKit sources
