+++ Packages/CmuxExtensionKit/Sources/CmuxExtensionKit/CMUXSidebarSnapshot.swift
@@ -1,0 +1,63 @@
+import Foundation
+
+public struct CMUXSidebarSnapshot: Codable, Equatable, Sendable {
+    public var apiVersion: CMUXExtensionAPIVersion
+    public var sequence: UInt64
+    public var windowID: UUID?
+    public var selectedWorkspaceID: UUID?
+    public var workspaces: [CMUXSidebarWorkspace]
+
+    public init(
+        apiVersion: CMUXExtensionAPIVersion = .sidebarV1,
+        sequence: UInt64,
+        windowID: UUID? = nil,
+        selectedWorkspaceID: UUID?,
+        workspaces: [CMUXSidebarWorkspace]
+    ) {
+        self.apiVersion = apiVersion
+        self.sequence = sequence
+        self.windowID = windowID
+        self.selectedWorkspaceID = selectedWorkspaceID
+        self.workspaces = workspaces
+    }
+}
+
+public struct CMUXSidebarWorkspace: Codable, Equatable, Identifiable, Sendable {
+    public var id: UUID
+    public var title: String
+    public var detail: String?
+    public var isPinned: Bool
+    public var rootPath: String?
+    public var projectRootPath: String?
+    public var gitBranch: String?
+    public var unreadCount: Int
+    public var latestNotification: String?
+    public var listeningPorts: [Int]
+    public var pullRequestURLs: [String]
+
+    public init(
+        id: UUID,
+        title: String,
+        detail: String? = nil,
+        isPinned: Bool = false,
+        rootPath: String? = nil,
+        projectRootPath: String? = nil,
+        gitBranch: String? = nil,
+        unreadCount: Int = 0,
+        latestNotification: String? = nil,
+        listeningPorts: [Int] = [],
+        pullRequestURLs: [String] = []
+    ) {
+        self.id = id
+        self.title = title
+        self.detail = detail
+        self.isPinned = isPinned
+        self.rootPath = rootPath
+        self.projectRootPath = projectRootPath
+        self.gitBranch = gitBranch
+        self.unreadCount = unreadCount
+        self.latestNotification = latestNotification
+        self.listeningPorts = listeningPorts
+        self.pullRequestURLs = pullRequestURLs
+    }
+}
